Add an attribute to control default action in attribute driven recipe
The kernel_modules.modules attribute allows to drive the default recipe
and configure/load/unload modules.
If you do not set an "action" the resource will be run the default one,
which is ":load".
With the new kernel_modules.default_module.action attribute, you can
change the default action for all modules handled in the attribute
driven recipe.

Change-Id: I1d7b121c6ab61dc1f355009171a553c9c2fa1c0e